[Mbstring][Php83][Php84] Isolate PHP 7 mb_* declarations into bootstrap72.php
The new mb_* polyfills (mb_str_pad, mb_ucfirst, mb_lcfirst, mb_trim,
mb_ltrim, mb_rtrim) trigger a warning and return false on PHP 7 in case
of invalid encoding, to avoid requiring polyfill-php80 transitively for
the \ValueError class. Their @return string|false annotation reflects
this, which contradicts the native PHP 8.4 contract.

Move the PHP 7 declarations into bootstrap72.php so the falsy return is
physically scoped to PHP 7, while bootstrap.php declares the PHP 8+
variants with : string typed return matching the native contract.

For Php83 and Php84, merge bootstrap80.php into bootstrap.php since
those files only use PHP 7.2-compatible syntax. Mbstring keeps its
bootstrap80.php because it uses PHP 8 union types.
