+++ src/Php84/bootstrap.php
@@ -49,34 +49,29 @@ if (!function_exists('fpow')) {
     function fpow(float $num, float $exponent): float { return p\Php84::fpow($num, $exponent); }
 }
 
-if (\PHP_VERSION_ID >= 80000) {
-    return require __DIR__.'/bootstrap80.php';
+if (\PHP_VERSION_ID < 80000) {
+    require __DIR__.'/bootstrap72.php';
 }
 
 if (extension_loaded('mbstring')) {
     if (!function_exists('mb_ucfirst')) {
-        /** @return string|false */
-        function mb_ucfirst(?string $string, ?string $encoding = null) { return p\Php84::mb_ucfirst((string) $string, $encoding); }
+        function mb_ucfirst(?string $string, ?string $encoding = null): string { return p\Php84::mb_ucfirst((string) $string, $encoding); }
     }
 
     if (!function_exists('mb_lcfirst')) {
-        /** @return string|false */
-        function mb_lcfirst(?string $string, ?string $encoding = null) { return p\Php84::mb_lcfirst((string) $string, $encoding); }
+        function mb_lcfirst(?string $string, ?string $encoding = null): string { return p\Php84::mb_lcfirst((string) $string, $encoding); }
     }
 
     if (!function_exists('mb_trim')) {
-        /** @return string|false */
-        function mb_trim(?string $string, ?string $characters = null, ?string $encoding = null) { return p\Php84::mb_trim((string) $string, $characters, $encoding); }
+        function mb_trim(?string $string, ?string $characters = null, ?string $encoding = null): string { return p\Php84::mb_trim((string) $string, $characters, $encoding); }
     }
 
     if (!function_exists('mb_ltrim')) {
-        /** @return string|false */
-        function mb_ltrim(?string $string, ?string $characters = null, ?string $encoding = null) { return p\Php84::mb_ltrim((string) $string, $characters, $encoding); }
+        function mb_ltrim(?string $string, ?string $characters = null, ?string $encoding = null): string { return p\Php84::mb_ltrim((string) $string, $characters, $encoding); }
     }
 
     if (!function_exists('mb_rtrim')) {
-        /** @return string|false */
-        function mb_rtrim(?string $string, ?string $characters = null, ?string $encoding = null) { return p\Php84::mb_rtrim((string) $string, $characters, $encoding); }
+        function mb_rtrim(?string $string, ?string $characters = null, ?string $encoding = null): string { return p\Php84::mb_rtrim((string) $string, $characters, $encoding); }
     }
 }
 
@@ -93,6 +88,10 @@ if (extension_loaded('bcmath')) {
     if (!function_exists('bcround')) {
         function bcround(string $num, int $precision = 0, $mode = RoundingMode::HalfAwayFromZero): string { return p\Php84::bcround($num, $precision, $mode); }
     }
+}
+
+if (\PHP_VERSION_ID >= 80200) {
+    return require __DIR__.'/bootstrap82.php';
 }
 
 if (extension_loaded('intl') && !function_exists('grapheme_str_split')) {
