+++ src/Util/TestListenerTrait.php
@@ -57,6 +57,8 @@ class TestListenerTrait
                 $bootstrap .= '81';
             } elseif (\PHP_VERSION_ID >= 80000 && file_exists($bootstrap.'80.php')) {
                 $bootstrap .= '80';
+            } elseif (\PHP_VERSION_ID < 80000 && file_exists($bootstrap.'72.php')) {
+                $bootstrap .= '72';
             }
             $bootstrap = new \SplFileObject($bootstrap.'.php');
             $newWarnings = 0;
